Update changed streams on PMT update
On playback, when a new version of the PMT (program map table) is received, start
updating the streams at the first stream that is changed, beginning with stream 0,
instead of resetting all streams.
This fixes playback problems that occur when there is a PMT version update in the
first part of the recording that is scanned before the playback starts.
At the start of the playback there is then a PMT change back to the previous version;
this change causes a reset of all stream data; the video playback is then started
with default size of 640x480 and default framerate of 29.97Hz.

Fixes #351
Fixes #trac-13557

Signed-off-by: Peter Bennett <pbennett@mythtv.org>
(cherry picked from commit c31db49cda268eaf7e822ffce4f1643100460d84)