mpegts-mythtv.c: pat_cb() harmonization part 3
add_program() with a following call to add_pid_to_program() is
functionally equivalent to add_pat_entry().

is_pat_same() is replaced by the equivalent check in the if statement
in the nested for loop.

This removes the MythTV addition Program.pid.

The local variable pmt_count is equivalent to ts->nb_prg.

pmt_pnums[i] is equivalent to ts->prg[i].id.
pmt_pids[i] is equivalent to ts->prg[i].pids[0] since the Program
is cleared, i.e. nb_pids is reset to 0 and then the pid added.

However, since the filters are created in the for loop and stored
in the ts->pids[] array, pmt_pids[i] is no longer needed.

The for loop is now unmodified.