mpegts-mythtv.c: export disposition from ISO_639_LANGUAGE_DESCRIPTOR
This replaces the use of the MythTV added, exported PMT for detecting
these in AvFormatDecoder::GetAudioTrackType(), which was causing
segmentation faults in a Finnish YLE sample I have with further
harmonization of mpegts-mythtv.c with FFmpeg's mpegts.c due to
a mismatch between the PMT stream numbers and FFmpeg's stream
numbers.

disposition is copied over to the AVStream in mpegts_add_stream().

language is only 4 bytes, so don't write past the end.
i.e. the loop is invalid.