mpegts-mythtv.c: harmonize write_section_data
FFmpeg inverted its if logic, so the second MythTV hunk is now following.
That hunk may not be necessary, since the issue did not exist upstream.
It may also be incorrect for the same reason.

This doesn't break MHEG but it doesn't start as early with one of
my samples, increasing from 5 seconds in to 18 seconds in.  I'm
not sure why this changed, however.