libavformat: add support for MPEG-PS and MPEG-TS changing streams in the middle of a file
This may be used to add and remove audio streams during a commercial break or
when a part time broadcaster switches from their pre-roll into a program.

Recordings from the Finnish broadcaster YLE have triggered this code.

ISO/IEC 13818-1:2021

For Program Streams:
§ 2.5.4 Program stream map
allows different version of the program stream map.

For Transport Streams:
§ 2.4.4.9 Program map table
also allows different versions of the program map table.

avprogram_id is an int to match type with AVProgram.  In mpegts.c,
SectionHeader::id is a uint16_t as ISO/IEC 13818-1:2021 specifies for program_number.

This allows us to ignore stream changes in programs that are not being watched
when there is more than one program.

An MPEG Program Stream can only have one program, so FFmpeg does not create an
AVProgram for it.  avprogram_id should be ignored in this case.  For example
av_find_program_from_stream() will always return NULL, allowing avprogram_id to
be easily ignored.