Revert "avformat/tls_openssl: add av_assert0() for tls_shared"
This reverts commits fd55c4b5f72a157fbb128d0ef203e9922b53552b
(for tls_openssl.c) and c74181a04b5f4e650eae662231e56518daef64d4
(for tls_gnutls.c).

It is impossible for a pointer to a member of a structure
to be NULL: If the containing structure exists, the member
exists and can't have a NULL address; if the containing
structure does not exist, then getting a pointer to the
substructure via &c->tls_shared would already be undefined
behavior.

Reviewed-by: Kacper Michajłow <kasper93@gmail.com>
Reviewed-by: Steven Liu <lingjiujianke@gmail.com>
Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>